+++ src/chameleon/__init__.py
@@ -3,3 +3,4 @@ from .zpt.template import PageTemplateFile
 from .zpt.template import PageTextTemplate
 from .zpt.template import PageTextTemplateFile
 from .zpt.loader import TemplateLoader as PageTemplateLoader
+from .exc import TemplateError