+++ src/chameleon/exc.py
@@ -99,12 +99,6 @@ def ellipsify(string, limit):
     return string
 
 
-def reconstruct_exc(cls, state):
-    exc = Exception.__new__(cls)
-    exc.__dict__ = state
-    return exc
-
-
 class TemplateError(Exception):
     """An error raised by Chameleon.
 
@@ -121,7 +115,7 @@ class TemplateError(Exception):
     And pickle/unpickled:
 
     >>> from pickle import dumps, loads
-    >>> loads(dumps(TemplateError(message, token)))
+    >>> loads(dumps(TemplateError(message, token), -1))
     TemplateError('message', 'token')
 
     """
@@ -130,23 +124,16 @@ class TemplateError(Exception):
         if not isinstance(token, Token):
             token = Token(token, 0)
 
-        self.msg = msg
-        self.token = safe_native(token)
-        self.offset = getattr(token, "pos", 0)
-        self.filename = token.filename
-        self.location = token.location
+        Exception.__init__(self, msg, token)
 
     def __copy__(self):
         inst = Exception.__new__(type(self))
-        inst.__dict__ = self.__dict__.copy()
+        inst.args = self.args
         return inst
 
-    def __reduce__(self):
-        return reconstruct_exc, (type(self), self.__dict__)
-
     def __str__(self):
-        text = "%s\n\n" % self.msg
-        text += " - String:     \"%s\"" % self.token
+        text = "%s\n\n" % self.args[0]
+        text += " - String:     \"%s\"" % safe_native(self.token)
 
         if self.filename:
             text += "\n"
@@ -161,10 +148,26 @@ class TemplateError(Exception):
     def __repr__(self):
         try:
             return "%s('%s', '%s')" % (
-                self.__class__.__name__, self.msg, self.token
+                self.__class__.__name__, self.args[0], safe_native(self.token)
                 )
         except AttributeError:
             return object.__repr__(self)
+
+    @property
+    def token(self):
+        return self.args[1]
+
+    @property
+    def filename(self):
+        return self.token.filename
+
+    @property
+    def location(self):
+        return self.token.location
+
+    @property
+    def offset(self):
+        return getattr(self.token, "pos", 0)
 
 
 class ParseError(TemplateError):